Rename dayPicker8TestUtils
